Delete old search users
