+++ backend/tpbackend/discord/commands/search_users.py
@@ -1,6 +1,7 @@
 from tpbackend.storage import User
 from .command import Command
-from tpbackend.utils import search_users, user_name
+from tpbackend.user.query import UserQuery
+from tpbackend.user.utils import md_user_link
 
 
 class SearchUsersCommand(Command):
@@ -19,26 +20,20 @@ Usage: `!search <query>`
         )
 
     def execute(self, user: User, msg: str) -> str:
-        results = search_users(msg, limit=50)
+        results = UserQuery.search(UserQuery.base(), msg).limit(10)
+        results = UserQuery.apply_sort(results, "name", "asc")
+        results = list(results)
         if len(results) == 0:
             return "No users found"
-        if self.is_admin(user):
-            return self.print_admin(results)
-        return self.print_regular(results)
+        return self.print(results=results, show_permissions=self.is_admin(user))
 
-    def print_regular(self, results: list[User]) -> str:
+    def print(self, results: list[User], show_permissions=False) -> str:
         out = ""
         for r in results:
-            out += f"- {r.id} - {user_name(r, as_markdown_link=True)}\n"
-            if len(out) > 1500:
-                return "Output too long. Narrow your search."
-        return out
-
-    def print_admin(self, results: list[User]) -> str:
-        out = ""
-        for r in results:
-            out += f"### {r.id} - {user_name(r, as_markdown_link=True)}\n"
-            out += f"Permissions: `{",".join(r.permissions)}`\n"
+            out += f"- {md_user_link(r)} ({r.get_id()})"
+            if show_permissions:
+                out += f" `{",".join(r.permissions)}`"
+            out += "\n"
             if len(out) > 1500:
                 return "Output too long. Narrow your search."
         return out
