+++ backend/tpbackend/globals.py
@@ -8,7 +8,7 @@ if DEBUG:
     print("⚠️⚠️⚠️ Running in DEBUG mode ⚠️⚠️⚠️")
 
 
-TIMEPLAYED_URL = os.environ.get("TIMEPLAYED_URL", "").rstrip("/")
+TIMEPLAYED_URL = os.environ.get("TIMEPLAYED_URL", "http://localhost").rstrip("/")
 
 MINIMUM_SESSION_LENGTH = int(os.environ.get("MINIMUM_SESSION_LENGTH", "0"))
